fix (client): array parameters should be name and array element pairs
All URI parameters should be one name and one value pair, but their can be multiple parameters with the same name.

We utilize this behavior for array values and add a name and value pair for each element in the array.

For example

'statuses[]': ["in progress", "completed"]

will become

statuses%5B%5D=in+progress&statuses%5B%5D=completed

FIXES: #21